[move] 評価関係のエンティティをhokutosai.server.data.entity.assessmentsに移動
