+++ src/main/java/hokutosai/server/data/entity/exhibitions/ExhibitionAssess.java
@@ -8,8 +8,8 @@ import javax.persistence.Table;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 
-import hokutosai.server.data.entity.Assess;
 import hokutosai.server.data.entity.account.SecureAccount;
+import hokutosai.server.data.entity.assessments.Assess;
 import hokutosai.server.error.InvalidParameterValueException;
 import lombok.Getter;
 import lombok.Setter;
