+++ src/main/java/hokutosai/server/data/json/shops/ShopAssessmentResponse.java
@@ -4,8 +4,8 @@ import lombok.Getter;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
-import hokutosai.server.data.entity.Assess;
-import hokutosai.server.data.entity.AssessedScore;
+import hokutosai.server.data.entity.assessments.Assess;
+import hokutosai.server.data.entity.assessments.AssessedScore;
 import hokutosai.server.data.json.AssessmentResponse;
 
 public class ShopAssessmentResponse extends AssessmentResponse {
